% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_apply_new_sample.R
\name{sim_apply_new_sample}
\alias{sim_apply_new_sample}
\title{Function to apply loadings from AC-PCA to new sample}
\usage{
sim_apply_new_sample(
  n = 5,
  b = 10,
  p = 400,
  alpha_orig = 2.5,
  alpha_new = 2.5,
  nsim = 100
)
}
\arguments{
\item{n}{the number of subjects; default is 5}

\item{b}{the number of brain regions; default is 10}

\item{p}{the number of features per brain region; default is 400}

\item{alpha_orig}{the constant that is multiplied by the confounder in the original simulated sample; default is 2.5}

\item{alpha_new}{the constant that is multiplied by the confounder in the new simulated sample; default is also 2.5}

\item{nsim}{number of simulations to run; default is 100}
}
\description{
User can specify alpha in the original sample and alpha in the new sample (i.e., samples differ in terms of strength of confounder)
}
