% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_compare_fun.R
\name{sim_compare_fun}
\alias{sim_compare_fun}
\title{Function to simulate data in the same way as in the 2016 AC-PCA paper and compare different methods (PCA, ComBat, SVA, AC-PCA)}
\usage{
sim_compare_fun(n = 5, b = 10, p = 400, alpha = 2.5, nsim = 100)
}
\arguments{
\item{n}{the number of subjects; default is 5}

\item{b}{the number of brain regions; default is 10}

\item{p}{the number of features per brain region; default is 400}

\item{alpha}{the constant that is multiplied by the confounder; default is 2.5}

\item{nsim}{number of simulations to run; default is 100}
}
\description{
Data will be simulated using sim_dat_fun function and then we will compare performance of ComBat, SVA, and acPCA, just like in the AC-PCA paper
}
